refactor(structure): refactor structure of architecture to create a MVC oriented design
