+++ src/main/kotlin/blackjack/model/Dealer.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.model
 
 import blackjack.view.InputView
 import blackjack.view.OutputView
