+++ src/main/kotlin/blackjack/model/Deck.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.model
 
 class Deck {
