+++ src/main/kotlin/blackjack/model/GameResult.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.model
 
 enum class GameResult(val label: String) {
     WIN("Win"),
