+++ src/main/kotlin/blackjack/model/Player.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.model
 
 class Player (name: String) : Participant(name) {
