+++ src/test/kotlin/blackjack/model/CardTest.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.model
 
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.Test
