+++ src/test/kotlin/blackjack/model/DeckTest.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.model
 
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.BeforeEach
