+++ src/test/kotlin/blackjack/model/HandTest.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.model
 
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
